+++ examples/simple_ze/main.c
@@ -143,33 +143,15 @@ tinytc_status_t custom_kernel(ze_context_handle_t context, ze_device_handle_t de
 
     CHECK(tinytc_ze_core_info_create(&info, device));
 
-    /*static const char source_text[] =*/
-        /*"func @copy(%A: memref<i32x" CHUNK_SIZE_S "x?>, %B: memref<i32x" CHUNK_SIZE_S "x?>) {\n"*/
-        /*"    %gid = builtin.group_id : index\n"*/
-        /*"    %a = subview %A[0:" CHUNK_SIZE_S ",%gid] : memref<i32x" CHUNK_SIZE_S ">\n"*/
-        /*"    %b = subview %B[0:" CHUNK_SIZE_S ",%gid] : memref<i32x" CHUNK_SIZE_S ">\n"*/
-        /*"    %c0 = constant 0 : i32\n"*/
-        /*"    %c1 = constant 1 : i32\n"*/
-        /*"    axpby.n %c1, %a, %c0, %b\n"*/
-        /*"}\n";*/
     static const char source_text[] =
-        "func @copy(%A: memref<f16x15x64> align(64), %B: memref<f16x64x64> align(64))\n"
-        "    subgroup_size(16) work_group_size(16,1) {\n"
-        "    %gid = builtin.group_id : index\n"
-        "    %0 = constant 0 : index\n"
-        "    %1 = cmp.eq %gid, %0 : bool\n"
-        "    if %1 {\n"
-        "       parallel {\n"
-        "           %lalilu = constant 0 : index\n"
-        "           %blilu = constant 0 : index\n"
-        "           %a = subview %A[0:15,0:64] : memref<f16x15x64,strided<1,15>>\n"
-        "           %2 = cooperative_matrix_load.n.both_checked %a[%lalilu,%blilu], align 64 : coopmatrix<f16x16x8,matrix_acc>\n"
-        "           cooperative_matrix_store %2, %B[%0,%0], align 64\n"
-        "       }\n"
-        "    }\n"
+        "func @copy(%A: memref<i32x" CHUNK_SIZE_S "x?>, %B: memref<i32x" CHUNK_SIZE_S "x?>) {\n"
+        "    %gid = builtin.group_id.x : index\n"
+        "    %a = subview %A[0:" CHUNK_SIZE_S ",%gid] : memref<i32x" CHUNK_SIZE_S ">\n"
+        "    %b = subview %B[0:" CHUNK_SIZE_S ",%gid] : memref<i32x" CHUNK_SIZE_S ">\n"
+        "    %c0 = constant 0 : i32\n"
+        "    %c1 = constant 1 : i32\n"
+        "    axpby.n %c1, %a, %c0, %b\n"
         "}\n";
-
-    printf("%s", source_text);
 
     CHECK(tinytc_parse_string(&program, sizeof(source_text), source_text, NULL));
     CHECK(tinytc_ze_kernel_bundle_create_with_program(&module, context, device, program, 0u));
@@ -177,7 +159,7 @@ tinytc_status_t custom_kernel(ze_context_handle_t context, ze_device_handle_t de
 
     ZE_CHECK(zeKernelSetArgumentValue(kernel, 0, sizeof(A), &A));
     ZE_CHECK(zeKernelSetArgumentValue(kernel, 1, sizeof(B), &B));
-    ze_group_count_t group_count = tinytc_ze_get_group_count(howmany);
+    ze_group_count_t group_count = {howmany, 1u, 1u};
     ZE_CHECK(zeCommandListAppendLaunchKernel(list, kernel, &group_count, NULL, 0, NULL));
     ZE_CHECK(zeCommandListHostSynchronize(list, TIMEOUT));
 
@@ -191,7 +173,9 @@ tinytc_status_t custom_kernel(ze_context_handle_t context, ze_device_handle_t de
         if (r < 16 && c < 8) {
             printf("%d ", host[i]);
         }
-        if (r == 63 && c < 8) { printf("\n"); }
+        if (r == 63 && c < 8) {
+            printf("\n");
+        }
         if (host[i] == i) {
             ++ok;
         }