+++ src/cl/kernel.cpp
@@ -128,10 +128,10 @@ tinytc_status_t tinytc_cl_get_group_size(cl_kernel kernel, size_t *local_size) {
                                  3 * sizeof(std::size_t), local_size, nullptr));
 }
 
-void tinytc_cl_get_global_size(int64_t howmany, const size_t *local_size, size_t *global_size) {
+void tinytc_cl_get_global_size(const size_t *num_groups, const size_t *local_size,
+                               size_t *global_size) {
     for (size_t i = 0; i < 3; ++i) {
-        global_size[i] = local_size[i];
+        global_size[i] = num_groups[i] * local_size[i];
     }
-    global_size[0] *= howmany;
 }
 }